Attempting fix for variant scoring function

+++ evo2_variant_scoring.py
@@ -72,7 +72,8 @@ def make_alt_seq(ref_seq: str, pos: int, ref: str, alt: str, slice_start: int):
 @torch.no_grad()
 def ll_score(model, tokenizer, seq: str):
     device = "cuda:0" if torch.cuda.is_available() else "cpu"
-    toks = torch.tensor(tokenizer(seq), dtype=torch.int16, device=device)[None, :]
+    # Change dtype from torch.int16 to torch.long for embedding compatibility
+    toks = torch.tensor(tokenizer(seq), dtype=torch.long, device=device)[None, :]
     out, _ = model(toks)
     logp = torch.log_softmax(out, -1)
     return logp.mean().item()
